backend/requirements.txt
fastapi>=0.115,<1
uvicorn>=0.34,<1
pydantic>=2.0,<3
python-dotenv>=1.0,<2
openai>=1.0,<2
langgraph>=0.2,<1
neo4j>=5.0,<6
